Added postgresql dependencies and functionality

+++ app.py
@@ -12,9 +12,12 @@ Authors:
 """
 
 from flask import Flask, Response, send_from_directory, render_template, request, url_for, redirect
-from os import environ, remove as rm
-from os.path import join, exists
+import os
+from os import environ
+from os.path import exists
 from sqlite3 import connect
+from init_db import init_db, db_connect
+import psycopg2
 import stripe
 
 # App Configuration
@@ -23,6 +26,14 @@ app = Flask(__name__,
             static_folder='static',
             template_folder='templates')
 
+# Checks for Heroku config var, sets to production mode if it finds DATABASE_URL is populated
+if os.environ.get('DATABASE_URL') is not None:
+    prod = True
+    param_query_symbol = '%s'
+else:
+    prod = False
+    param_query_symbol = '?'
+    
 # Name of the database file
 DB = 'database.db'
 
@@ -30,38 +41,60 @@ DB = 'database.db'
 stripe_key = 'sk_test_51MZ2KAClce1MywlhOsuuW61sleJa4FX39mSt8bQbmBIGb6i2PVf4jAideajXjKTWUENOjq7jxijtWOWVwtBlDC2q00PPk1A193'
 
 
-# Set up database if it doesn't exist
+# Set up local database if it doesn't exist
 def create_db():
-    with connect(DB) as conn:
-        cur = conn.cursor()
-        cur.executescript("""
-            DROP TABLE IF EXISTS customer;
-            DROP TABLE IF EXISTS service;
-            CREATE TABLE customer(
-                email TEXT NOT NULL PRIMARY KEY,
-                first_name TEXT NOT NULL,
-                last_name TEXT NOT NULL,
-                phone TEXT NOT NULL
-            );
-            CREATE TABLE service(
-                service_id INTEGER PRIMARY KEY,
-                service_type TEXT NOT NULL,
-                date DATE NOT NULL,
-                time TEXT NOT NULL,
-                completed int NOT NULL,
-                balance REAL NOT NULL,
-                paid int NOT NULL,
-                customer_email INTEGER NOT NULL,
-                FOREIGN KEY(customer_email) REFERENCES customer(email)
-            );
-        """)
+    if not prod:
+        with connect(DB) as conn:
+            cur = conn.cursor()
+            cur.executescript("""
+                DROP TABLE IF EXISTS customer;
+                DROP TABLE IF EXISTS service;
+                CREATE TABLE customer(
+                    email TEXT NOT NULL PRIMARY KEY,
+                    first_name TEXT NOT NULL,
+                    last_name TEXT NOT NULL,
+                    phone TEXT NOT NULL
+                );
+                CREATE TABLE service(
+                    service_id INTEGER PRIMARY KEY,
+                    service_type TEXT NOT NULL,
+                    date DATE NOT NULL,
+                    time TEXT NOT NULL,
+                    completed int NOT NULL,
+                    balance REAL NOT NULL,
+                    paid int NOT NULL,
+                    customer_email INTEGER NOT NULL,
+                    FOREIGN KEY(customer_email) REFERENCES customer(email)
+                );
+            """)
+    else:
+        try:
+            conn = db_connect()
+            cur = conn.cursor()
+            cur.execute('SELECT * FROM customer')
+        except Exception as e:
+            print('exception in init db')
+            init_db()
+            return redirect('index.html')
+        
 
 
 # Check if a customer exists
 def customer_exists(customer_email):
-    with connect(DB) as conn:
-        cur = conn.cursor()
-        customer = cur.execute("SELECT * FROM customer WHERE email = ?", (customer_email,)).fetchone()
+    if not prod:
+        conn = connect(DB)
+    else:
+        conn = db_connect()
+
+    cur = conn.cursor()
+
+    cur.execute("SELECT * FROM customer WHERE email = {0}".format(param_query_symbol), (customer_email,))
+    
+    customer = cur.fetchone()
+
+    conn.commit()
+    conn.close()
+
     if customer is not None:
         return True
     else:
@@ -79,31 +112,47 @@ def main_view():
 def schedule_view():
     # If posting a web form to schedule service:
     if request.method == "POST":
-        with connect(DB) as conn:
-            cur = conn.cursor()
+        if not prod:
+            conn = connect(DB)
+            
+        else:
+            conn = db_connect()
 
-            # Check if customer exists, create them if not
-            if not customer_exists(request.form['email']):
-                cur.execute("INSERT INTO customer (first_name, last_name, email, phone) VALUES (?,?,?,?)", (request.form['first_name'], request.form['last_name'], request.form['email'], request.form['phone_number']))
+        cur = conn.cursor()
 
-            # Create the service
-            cur.execute("INSERT INTO service (service_type, date, time, completed, balance, paid, customer_email) VALUES (?,?,?,?,?,?,?)", (request.form['service_type'], request.form['date'], request.form['time'], '0', '0.00', '0', request.form['email']))
+        # Check if customer exists, create them if not
+        if not customer_exists(request.form['email']):
+            cur.execute("INSERT INTO customer (first_name, last_name, email, phone) VALUES ({0},{0},{0},{0})".format(param_query_symbol), (request.form['first_name'], request.form['last_name'], request.form['email'], request.form['phone_number']))
+
+        # Create the service
+        cur.execute("INSERT INTO service (service_type, date, time, completed, balance, paid, customer_email) VALUES ({0},{0},{0},{0},{0},{0},{0})".format(param_query_symbol), (request.form['service_type'], request.form['date'], request.form['time'], '0', '0.00', '0', request.form['email']))
+        
+        conn.commit()
+        conn.close()
 
         return redirect(url_for('schedule_view'))
     # Regular GET request
     else:
-        with connect(DB) as conn:
-            cur = conn.cursor()
-            results = cur.execute("SELECT * FROM customer").fetchall()
-        return render_template('schedule.html', results=results)
+        return render_template('schedule.html')
 
 
 # Admin view
 @app.route('/joeazzi')
 def admin_view():
-    with connect(DB) as conn:
-        cur = conn.cursor()
-        services = cur.execute("SELECT service_id, service_type, date, time, first_name, last_name FROM service INNER JOIN customer on customer.email=service.customer_email").fetchall()
+    if not prod:
+        conn = connect(DB)
+        
+    else:
+        conn = db_connect()
+
+    cur = conn.cursor()
+
+    cur.execute("SELECT service_id, service_type, date, time, first_name, last_name FROM service INNER JOIN customer on customer.email=service.customer_email")
+    services = cur.fetchall()
+
+    conn.commit()
+    conn.close()
+
     return render_template('admin.html', services=services)
 
 
@@ -112,19 +161,49 @@ def admin_view():
 def service_ticket_view():
     ticket_num = request.args.get('ticket')
     if request.method == "POST":
-        with connect(DB) as conn:
-            cur = conn.cursor()
-            if request.form['balance'] == '':
-                balance_tuple = cur.execute("SELECT balance FROM service WHERE service_id = ?", (ticket_num,)).fetchone()
-                balance = balance_tuple[0]
-            else:
-                balance = request.form['balance']
-            cur.execute("UPDATE service SET service_type = ?, date = ?, time = ?, completed = ?, balance = ?, paid = ? WHERE service_id = ?", (request.form['service_type'], request.form['date'], request.form['time'], int(request.form['completed']), balance, int(request.form['paid']), ticket_num,))
-            return redirect('/joeazzi/service?ticket={}'.format(ticket_num))
+        if not prod:
+            conn = connect(DB)
+        else:
+            conn = db_connect()
+        cur = conn.cursor()
+        if request.form['balance'] == '':
+            cur.execute("SELECT balance FROM service WHERE service_id = {0}".format(param_query_symbol), (ticket_num,))
+            balance_tuple = cur.fetchone()
+            balance = balance_tuple[0]
+        else:
+            balance = request.form['balance']
+
+        if request.form['date'] == '':
+            cur.execute('SELECT date FROM service WHERE service_id = {}'.format(ticket_num))
+            date = cur.fetchone()
+        else:
+            date = request.form['date']
+
+        if request.form['time'] == '':
+            cur.execute('SELECT time FROM service WHERE service_id = {}'.format(ticket_num))
+            time = cur.fetchone()
+        else:
+            time = request.form['time']
+
+        cur.execute("UPDATE service SET service_type = {0}, date = {0}, time = {0}, completed = {0}, balance = {0}, paid = {0} WHERE service_id = {0}".format(param_query_symbol), (request.form['service_type'], date, time, int(request.form['completed']), balance, int(request.form['paid']), ticket_num,))
+
+        conn.commit()
+        conn.close()
+
+        return redirect('/joeazzi/service?ticket={}'.format(ticket_num))
     else:
-        with connect(DB) as conn:
-            cur = conn.cursor()
-            service = cur.execute("SELECT * FROM service INNER JOIN customer on customer.email=service.customer_email WHERE service.service_id = ?", (ticket_num,)).fetchone()
+        if not prod:
+            conn = connect(DB)
+        else:
+            conn = db_connect()
+        
+        cur = conn.cursor()
+        cur.execute("SELECT * FROM service INNER JOIN customer on customer.email=service.customer_email WHERE service.service_id = {0}".format(param_query_symbol), (ticket_num,))
+        service = cur.fetchone()
+
+        conn.commit()
+        conn.close()
+
         return render_template('service-ticket.html', service=service)
 
 
@@ -132,17 +211,23 @@ def service_ticket_view():
 @app.route('/joeazzi/service/delete')
 def delete_service_ticket():
     ticket_num = request.args.get('ticket')
-    with connect(DB) as conn:
-        cur = conn.cursor()
-        cur.execute("DELETE FROM service WHERE service_id = ?", (ticket_num,))
+    if not prod:
+        conn = connect(DB)
+    else:
+        conn = db_connect()
+        
+    cur = conn.cursor()
+    cur.execute("DELETE FROM service WHERE service_id = {0}".format(param_query_symbol), (ticket_num,))
+
+    conn.commit()
+    conn.close()
+
     return redirect('/joeazzi')
 
 
 # Main calling function
 if __name__ == '__main__':
-
-    # Create the database if it doesn't exist
-    if not exists(DB):
+    if not exists(DB) or prod:
         create_db()
 
     # App configuration
